// modules/dashboard/certificate/editor/interfaces/Editor.ts
export enum EditorPanel {
  Settings = 'settings-panel',
  Text = 'text-panel',
  Elements = 'elements-panel',
  Image = 'image-panel',
  Input = 'input-panel',
  ImageProperties = 'image-properties-panel',
  TextProperties = 'text-properties-panel',
  RectangleProperties = 'rectangle-properties-panel',
  LineProperties = 'line-properties-panel',
  InputProperties = 'input-properties-panel',
}